Combined event.model and event.schema into one reusable file.
